refactor(backoffice-v2): moved useTransactionsLogic to a file
