+++ apps/backoffice-v2/src/pages/TransactionMonitoringAlerts/TransactionMonitoringAlerts.page.tsx
@@ -1,64 +1,8 @@
-import { useAlertsQuery } from '@/domains/alerts/hooks/queries/useAlertsQuery/useAlertsQuery';
 import { AlertAnalysisSheet } from '@/pages/TransactionMonitoringAlerts/components/AlertAnalysisSheet';
 import { AlertsTable } from '@/pages/TransactionMonitoringAlerts/components/AlertsTable';
 import { AlertsHeader } from 'src/pages/TransactionMonitoringAlerts/components/AlertsHeader';
-import { useToggle } from '@/common/hooks/useToggle/useToggle';
-import { useZodSearchParams } from '@/common/hooks/useZodSearchParams/useZodSearchParams';
-import { useAuthenticatedUserQuery } from '@/domains/auth/hooks/queries/useAuthenticatedUserQuery/useAuthenticatedUserQuery';
-import { useUsersQuery } from '@/domains/users/hooks/queries/useUsersQuery/useUsersQuery';
-import { useMemo } from 'react';
-import { usePagination } from '@/common/hooks/usePagination/usePagination';
-import { getAlertsSearchSchema } from '@/pages/TransactionMonitoringAlerts/get-alerts-search-schema';
 import { AlertsPagination } from '@/pages/TransactionMonitoringAlerts/AlertsPagination/AlertsPagination';
-import { useSearch } from '@/common/hooks/useSearch/useSearch';
-
-export const useTransactionMonitoringAlertsLogic = () => {
-  const { data: session } = useAuthenticatedUserQuery();
-  const AlertsSearchSchema = getAlertsSearchSchema(session?.user?.id);
-  const [{ filter, sortBy, sortDir, page, pageSize, search: searchValue }] =
-    useZodSearchParams(AlertsSearchSchema);
-  const { data: alerts } = useAlertsQuery({
-    filter,
-    page,
-    pageSize,
-    search: searchValue,
-    sortDir,
-    sortBy,
-  });
-  const { data: assignees } = useUsersQuery();
-  const sortedAssignees = useMemo(
-    () =>
-      // Sort assignees so that the authenticated user is always first
-      assignees
-        ?.slice()
-        ?.sort((a, b) => (a?.id === session?.user?.id ? -1 : b?.id === session?.user?.id ? 1 : 0)),
-    [assignees, session?.user?.id],
-  );
-  const [isSheetOpen, toggleIsAlertAnalysisSheetOpen, toggleOnIsAlertAnalysisSheetOpen] =
-    useToggle();
-  const { onPaginate, onPrevPage, onNextPage } = usePagination();
-  const isLastPage = (alerts?.length ?? 0 < pageSize) || alerts?.length === 0;
-  const { search, onSearch } = useSearch({
-    initialSearch: searchValue,
-  });
-
-  return {
-    alerts,
-    isSheetOpen,
-    toggleOnIsAlertAnalysisSheetOpen,
-    toggleIsAlertAnalysisSheetOpen,
-    assignees: sortedAssignees,
-    authenticatedUserId: session?.user?.id,
-    page,
-    pageSize,
-    onPrevPage,
-    onNextPage,
-    onPaginate,
-    isLastPage,
-    search,
-    onSearch,
-  };
-};
+import { useTransactionMonitoringAlertsLogic } from '@/pages/TransactionMonitoringAlerts/hooks/useTransactionMonitoringAlertsLogic/useTransactionMonitoringAlertsLogic';
 
 export const TransactionMonitoringAlerts = () => {
   const {
